+++ src/commonMain/kotlin/csense/kotlin/general.kt
@@ -1,5 +1,4 @@
 @file:OptIn(ExperimentalContracts::class)
-@file:Suppress("Annotator")
 
 package csense.kotlin